Gracefully handle dead stat pipes.
The JSON stat server will now timeout if it can't read from a pipe and
unlink the pipe.

The unlinking is kind of heavy-handed and weird for the server to be
doing, but since procs can be KILLed and leave behind their pipes,
something has to clean them up. A `pipe_cleaner` process would probably
be best, and also has a cool name. The future!